Done with Skills and Project Routes

+++ controllers/projectcontroller/deleteProject.js
@@ -5,25 +5,22 @@ const deleteproject = async (req, res) => {
     const project = await ProjectModel.findById(req.params.id);
 
     if (!project) {
-      return res.status(404).json({ message: 'Skill not found' });
+      return res.status(404).json({ 
+        success:false,
+        message: 'Project not found' 
+      });
     }
 
-    // Delete the document from MongoDB.
-    await ProjectModel.findByIdAndRemove(req.params.id);
-
-    const imagePath = `../../client/src/images/uploads/projects/${project.image}`;
-
-    fs.unlink(imagePath, (err) => {
-      if (err) {
-        console.error('Error deleting image file:', err);
-      }
-    });
-
-    res.status(200).json({ message: 'Project deleted successfully' });
+    const result = await cloudinary.uploader.destroy(project.publicid);
+    if (result.result==='ok') {
+      await ProjectModel.findByIdAndRemove(req.params.id);
+      res.status(200).json({success:true,
+        message: 'Project deleted successfully' });
+    }
   } catch (error) {
     console.error('Error deleting project:', error);
     res.status(500).json({ error: 'Internal server error' });
   }
 };
 
-export default deleteskill;
+export default deleteproject;
